+++ src/Internals/Details/Info.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Walletable\Wallet\Details;
+namespace Walletable\Internals\Details;
 
 use Exception;
