+++ src/Internals/Details/MoneyCast.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Walletable\Wallet\Details;
+namespace Walletable\Internals\Details;
 
 use Walletable\Money\Money;
