+++ src/Internals/Lockers/OptimisticLocker.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Walletable\Lockers;
+namespace Walletable\Internals\Lockers;
 
 use Illuminate\Support\Facades\DB;
 use Walletable\Models\Transaction;
